requirements.txt
imbalanced_learn==0.12.3
joblib==1.4.2
matplotlib==3.8.4
numpy==2.2.2
pandas==2.2.3
Pillow==11.1.0
plotly==5.22.0
scikit_learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
streamlit==1.32.0
